Documentation and package layout
changed the system to have a package layout, and added docstring for the countryPickerWindow class.
